Change int16_t to int to improve speed

+++ brute_force.cpp
@@ -2,13 +2,13 @@
 #include "char_compress.hpp"
 
 int brute_force_search(
-    const char *text, int text_length, const char *pattern, int16_t pattern_length,
-    int *output, int max_output_cnt, int16_t *fail
+    const char *text, int text_length, const char *pattern, int pattern_length,
+    int *output, int max_output_cnt, int *fail
 ) {
     int output_cnt = 0;
     for (int i = 0; i <= text_length - pattern_length; i++) {
         bool match = true;
-        for (int16_t j = 0; j < pattern_length; j++) {
+        for (int j = 0; j < pattern_length; j++) {
             if (get(text, i+j) != get(pattern, j)) {
                 match = false;
                 break;
