+++ state_machine_async_copy.cuh
@@ -1,6 +1,6 @@
 #pragma once
 
 int state_machine_search_async_copy(
-    const char *text, int text_length, const char *pattern, int16_t pattern_length,
-    int *output, int max_output_cnt, int16_t *fail
+    const char *text, int text_length, const char *pattern, int pattern_length,
+    int *output, int max_output_cnt, int *fail
 );
