+++ state_machine_unified.cuh
@@ -1,5 +1,5 @@
 #pragma once
 
 int state_machine_search_unified(
-    const char *text, int text_length, const char *pattern, int16_t pattern_length,
-    int *output, int max_output_cnt, int16_t *fail);
+    const char *text, int text_length, const char *pattern, int pattern_length,
+    int *output, int max_output_cnt, int *fail);
